// generated-contracts/models/call-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DeFi API (API users)
 * API surface available via API key authentication
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CallDto
 */
export interface CallDto {
    /**
     * Target contract address
     * @type {string}
     * @memberof CallDto
     */
    to: string;
    /**
     * ETH value in wei
     * @type {string}
     * @memberof CallDto
     */
    value: string;
    /**
     * Encoded function call data
     * @type {string}
     * @memberof CallDto
     */
    data: string;
}

/**
 * Check if a given object implements the CallDto interface.
 */
export function instanceOfCallDto(value: object): value is CallDto {
    if (!('to' in value) || value['to'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function CallDtoFromJSON(json: any): CallDto {
    return CallDtoFromJSONTyped(json, false);
}

export function CallDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CallDto {
    if (json == null) {
        return json;
    }
    return {
        
        'to': json['to'],
        'value': json['value'],
        'data': json['data'],
    };
}

export function CallDtoToJSON(json: any): CallDto {
    return CallDtoToJSONTyped(json, false);
}

export function CallDtoToJSONTyped(value?: CallDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'to': value['to'],
        'value': value['value'],
        'data': value['data'],
    };
}